refactor: migrate tool assignment logic to ToolControllerModern
